requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.6.0
pytest==7.3.0
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
railway==0.0.4
sqlparse==0.4.3
